fix: refactor: reorganize folder structure and update router paths
