--- src/pages/requests/MyRequestsPage.jsx
@@ -1,6 +1,0 @@
-// ✅ Employee/Manager/Admin - Ver/crear MIS solicitudes
-// MyRequestsPage.jsx contiene:
-// - Calendario
-// - Formulario nueva solicitud
-// - Balance
-// - Tabla de mis solicitudes
