--- src/pages/requests/TeamRequestsPage.jsx
@@ -1,1 +1,0 @@
-// ✅ Manager/Admin - Aprobar/rechazar solicitudes del equipo
